+++ tuplities-derive/src/lib.rs
@@ -10,6 +10,7 @@ mod tuplities_debug;
 mod tuplities_default;
 mod tuplities_eq;
 mod tuplities_first_tuple_row;
+mod tuplities_flatten_nested_tuple;
 mod tuplities_from;
 mod tuplities_hash;
 mod tuplities_index;
@@ -18,6 +19,7 @@ mod tuplities_last_tuple_row;
 mod tuplities_len;
 mod tuplities_mut;
 mod tuplities_mut_map;
+mod tuplities_nest_tuple;
 mod tuplities_option;
 mod tuplities_ord;
 mod tuplities_partial_eq;
@@ -285,10 +287,26 @@ pub fn impl_first_tuple_row(_attr: TokenStream, item: TokenStream) -> TokenStrea
     item.into()
 }
 
+/// Generate `FlattenNestedTuple` trait implementations for all tuple sizes.
+#[proc_macro_attribute]
+pub fn impl_flatten_nested_tuple(_attr: TokenStream, item: TokenStream) -> TokenStream {
+    let mut item = proc_macro2::TokenStream::from(item);
+    item.extend(tuplities_flatten_nested_tuple::impl_flatten_nested_tuple());
+    item.into()
+}
+
 /// Generate `LastTupleRow` trait implementations for all tuple sizes.
 #[proc_macro_attribute]
 pub fn impl_last_tuple_row(_attr: TokenStream, item: TokenStream) -> TokenStream {
     let mut item = proc_macro2::TokenStream::from(item);
     item.extend(tuplities_last_tuple_row::impl_last_tuple_row());
+    item.into()
+}
+
+/// Generate `NestTuple` trait implementations for all tuple sizes.
+#[proc_macro_attribute]
+pub fn impl_nest_tuple(_attr: TokenStream, item: TokenStream) -> TokenStream {
+    let mut item = proc_macro2::TokenStream::from(item);
+    item.extend(tuplities_nest_tuple::impl_nest_tuple());
     item.into()
 }